Update patch set 1
Patch Set 1: Code-Review-1

I think that sonar error is not correct because in that exception block, the exception is in fact logged.  The exception is logged and not re-raised intentionally because that class of error should not be considered fatal hence the log priority being warning.

Could we understand why sonar is complaining instead?

Patch-set: 1
Label: Code-Review=-1