Update patch set 1
Patch Set 1:

Without the change I'm suggesting, the wild card match for typeName will fail - you can try it yourself and see. If you do GET "/dcae-service-types?typeName=somepartialtypename*" it will not work; it was working before your change.
The reason for this I think, is you have to call resolveTypeName(typeName) AFTER the  DcaeServiceTypesQueryStatement.create() method. 
Since this method sets "type_name LIKE :typeName" if typeName.contains("*"); and then resolveTypeName() changes * to % by doing: replaceAll("\\*", "%")
If you don't follow this sequence and replace * to % before calling DcaeServiceTypesQueryStatement.create() then we won't do "type_name LIKE :typeName" so we loose the LIKE when we have wildcard match.
As for typeName is used in line 122 and 172; this might be an issue if we're using wild card match and we have results more than the default offset which is not likely since we're matching against partial typeName and not getting all dcae-service-tpyes; you could however, call again the resolveTypeName() before these lines but we never had any issues w/ this.

Patch-set: 1